Fixed Create Page
I Fixed the issue with the description box, I changed it from "input" to "textarea", that fixed the text wrapping issue I was facing last week. Also, I've noticed that, after posting a new card, the home page would be open, so I fixed this issue as I changed the opened page, so the "explore" page would be the page opened new the user create a new post.